solutions/agent-memory/pyproject.toml
[project]
name = "agent-memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.39.11",
    "cognee[neptune]==0.2.3",
    "mem0ai[extras, graph]>=0.1.116",
    "opensearch-py>=3.0.0",
    "python-dotenv>=1.1.1",
    "graphiti_core",
    "streamlit>=1.48.1",
    "strands-agents>=1.5.0",
    "strands-agents-tools>=0.2.4",
    "ddgs>=9.5.4",
]

[tool.uv.sources]
graphiti-core = { path = "graphiti_core-0.18.5-py3-none-any.whl" }

[tool.uv]
override-dependencies = [
    "fsspec==2025.7.0",
    "langchain-neo4j==0.5.0"
]
